OptimalBranchingCore

Documentation for the core module OptimalBranchingCore.jl, which provides the core functionality for the optimal branching algorithm.

API

OptimalBranchingCore.AbstractBranchingStrategy

OptimalBranchingCore.AbstractMeasure

OptimalBranchingCore.AbstractProblem

OptimalBranchingCore.AbstractPruner

OptimalBranchingCore.AbstractReducer

OptimalBranchingCore.AbstractResult

OptimalBranchingCore.AbstractSelector

OptimalBranchingCore.AbstractSetCoverSolver

OptimalBranchingCore.AbstractTableSolver

OptimalBranchingCore.Branch

OptimalBranchingCore.BranchingTable

OptimalBranchingCore.Clause

OptimalBranchingCore.DNF

OptimalBranchingCore.IPSolver

OptimalBranchingCore.LPSolver

OptimalBranchingCore.NoProblem

OptimalBranchingCore.NoPruner

OptimalBranchingCore.NoResult

OptimalBranchingCore.OptBranchingStrategy

OptimalBranchingCore.SolverConfig

OptimalBranchingCore.SubCover

OptimalBranchingCore.apply

OptimalBranchingCore.branch

OptimalBranchingCore.complexity

OptimalBranchingCore.cover

OptimalBranchingCore.cover

OptimalBranchingCore.cover

OptimalBranchingCore.covered_by

OptimalBranchingCore.covered_by

OptimalBranchingCore.measure

OptimalBranchingCore.optimal_branching

OptimalBranchingCore.problem_reduce

OptimalBranchingCore.prune

OptimalBranchingCore.result

OptimalBranchingCore.select

OptimalBranchingCore.solve_branches

OptimalBranchingCore.solve_table

OptimalBranchingCore.subcovers

OptimalBranchingCore.subcovers

OptimalBranchingCore.apply — Method

apply(::NoProblem, ::Clause, vs)

Applies a clause to a NoProblem instance, returning a NoProblem instance. This function serves as a placeholder for scenarios where no valid problem is present.

Arguments

::NoProblem: An instance of NoProblem, representing the absence of a problem.

::Clause: A clause that is being applied (not used in this context).

vs: A vector of variables (not used in this context).

Returns

NoProblem(): An instance of NoProblem, indicating that no problem exists.

OptimalBranchingCore.branch — Method

Branch the given problem using the specified solver configuration.

# Arguments
- `p::P`: The problem instance to branch.
- `config::SolverConfig`: The configuration for the solver.

# Returns
The maximum result obtained from the branches.

OptimalBranchingCore.complexity — Method

complexity(sbranches::Vector{T}) where {T}

Calculates the complexity based on the provided branching structures.

Arguments

sbranches::Vector{T}: A vector of branching structures, where each element represents a branch.

Returns

Float64: The computed complexity value.

OptimalBranchingCore.cover — Method

cover(sub_covers::AbstractVector{SubCover{INT}}, dns::Vector{TF}, solver::IPSolver; verbose::Bool = false) where{INT, TF}

This function implements a cover selection algorithm using an iterative process. It utilizes an integer programming solver to optimize the selection of sub-covers based on their complexity.

Arguments

sub_covers::AbstractVector{SubCover{INT}}: A vector of sub-cover objects that represent the available covers.

dns::Vector{TF}: A vector of decision variables or parameters that influence the complexity calculation.

solver::IPSolver: An object that contains the settings for the integer programming solver, including the maximum number of iterations.

verbose::Bool: A flag to control the verbosity of the output. If set to true, additional information will be logged.

Returns

A tuple containing:

picked: A vector of selected sub-covers based on the optimization process.

cx: The final complexity value after the optimization iterations.

Description

This function iteratively solves the integer programming problem to find the optimal cover, updating the complexity value until convergence or the maximum number of iterations is reached.

OptimalBranchingCore.cover — Method

cover(sub_covers::AbstractVector{SubCover{INT}}, dns::Vector{TF}, solver::LPSolver; verbose::Bool = false) where{INT, TF}

Finds the optimal cover based on the provided differences in measure.

Arguments

sub_covers::AbstractVector{SubCover{INT}}: A vector of subcover structures.

dns::Vector{TF}: A vector of differences in measure for each subcover.

solver::LPSolver: The linear programming solver to be used.

verbose::Bool: A flag to enable verbose output (default is false).

Returns

A tuple containing:

A vector of selected subcovers.

The computed complexity value.

Description

This function iteratively solves the linear programming problem to find the optimal cover, updating the complexity value until convergence or the maximum number of iterations is reached.

OptimalBranchingCore.cover — Method

cover(sub_covers::AbstractVector{SubCover{INT}}, p::P, m::M, vs::Vector{T}, solver::Union{LPSolver, IPSolver}; verbose::Bool = false) where{INT, P<:AbstractProblem, M<:AbstractMeasure, T}

Calculates the optimal cover from the provided subcovers using a specified solver.

Arguments

sub_covers::AbstractVector{SubCover{INT}}: A vector of subcover structures.

p::P: An instance of a problem that needs to be solved.

m::M: An instance of a measure associated with the problem.

vs::Vector{T}: A vector of values used in the calculation.

solver::Union{LPSolver, IPSolver}: The solver to be used for optimization.

verbose::Bool: A flag to enable verbose output (default is false).

Returns

A tuple containing:

A vector of selected subcovers.

The computed complexity value.

Description

This function computes the difference in measure for each subcover and then calls another cover function to find the optimal cover based on the computed differences.

OptimalBranchingCore.covered_by — Method

covered_by(a::LongLongUInt, b::LongLongUInt, mask::LongLongUInt)

Check if a is covered by b with mask. The function returns true if and only if a and b are the same when masked by mask.

OptimalBranchingCore.covered_by — Method

covered_by(a::LongLongUInt, clause::Clause)

Check if a is covered by the clause. The function returns true if and only if a and clause.val are the same when masked by clause.mask.

OptimalBranchingCore.measure — Method

measure(::NoProblem, ::AbstractMeasure)

Calculates a measure for a NoProblem instance. This function serves as a placeholder for scenarios where no valid problem is present.

Arguments

::NoProblem: An instance of NoProblem, representing the absence of a problem.

::AbstractMeasure: An abstract measure type, which is not utilized in this context.

Returns

Int: The measure value, which is always 0 for a NoProblem instance.

OptimalBranchingCore.optimal_branching — Method

optimal_branching(tbl::BranchingTable{INT}, vs::Vector{T}, problem::P, measure::M, solver::S, ::Type{R}; verbose::Bool = false) where{INT, T, P<:AbstractProblem, M<:AbstractMeasure, S<:AbstractSetCoverSolver, R<:AbstractResult}

Generate optimal branches from a given branching table.

Arguments

tbl::BranchingTable{INT}: The branching table containing subcovers.

vs::Vector{T}: A vector of variables to be used in the branching.

problem::P: The problem instance being solved.

measure::M: The measure used for evaluating the branches.

solver::S: The solver used for the set cover problem.

::Type{R}: The type of the result expected.

verbose::Bool: Optional; if true, enables verbose output (default is false).

Returns

A vector of Branch objects representing the optimal branches derived from the subcovers.

OptimalBranchingCore.problem_reduce — Method

problem_reduce(p::NoProblem, ::AbstractReducer, ::Type{R}) where{R<:AbstractResult}

Reduces a problem represented by a NoProblem instance. This function serves as a placeholder for scenarios where no valid problem is present.

Arguments

p::NoProblem: An instance of NoProblem, representing the absence of a problem.

::AbstractReducer: An abstract reducer type, which is not utilized in this context.

::Type{R}: The type of result expected, which must be a subtype of AbstractResult.

Returns

NoProblem(): An instance of NoProblem, indicating that no problem exists.

OptimalBranchingCore.prune — Method

prune(bt::BranchingTable, ::NoPruner, ::M, ::P, vs)

Applies a no-operation pruning strategy to the given branching table. This function serves as a placeholder for scenarios where no pruning is required.

Arguments

bt::BranchingTable: The branching table to be pruned.

::NoPruner: An instance of NoPruner, indicating that no pruning will occur.

::M: An abstract measure type, which is not utilized in this context.

::P: An abstract problem type, which is not utilized in this context.

vs: A vector of values associated with the branching process.

Returns

bt: The original branching table, unchanged.

OptimalBranchingCore.result — Method

result(::NoProblem, ::Clause, vs, ::Type{R}) where{R<:AbstractResult}

Generates a result from a NoProblem instance when a clause is applied. This function serves as a placeholder for scenarios where no valid result is present.

Arguments

::NoProblem: An instance of NoProblem, representing the absence of a problem.

::Clause: A clause that is being applied (not used in this context).

vs: A vector of variables (not used in this context).

::Type{R}: The type of result expected, which must be a subtype of AbstractResult.

Returns

NoResult(): An instance of NoResult, indicating that no valid result exists.

OptimalBranchingCore.select — Method

select(::NoProblem, ::AbstractMeasure, ::NoSelector)

Selects a branching strategy for a NoProblem instance. This function serves as a placeholder for scenarios where no valid selection is present.

Arguments

::NoProblem: An instance of NoProblem, representing the absence of a problem.

::AbstractMeasure: An abstract measure type, which is not utilized in this context.

::NoSelector: An instance of NoSelector, representing the absence of a selector.

Returns

nothing: Indicates that no selection is made due to the absence of a problem.

OptimalBranchingCore.solve_branches — Method

Solve branches of the given problem using the specified branching strategy.

# Arguments
- `p::P`: The problem instance to solve branches for.
- `strategy::OptBranchingStrategy`: The strategy to use for branching.
- `result_type::Type{R}`: The type of the result expected.

# Returns
A vector of branches derived from the problem using the specified strategy.

OptimalBranchingCore.solve_table — Method

solve_table

Solves a given problem using a specified table solver.

Arguments

::NoProblem: An instance of NoProblem, representing the absence of a problem.

::AbstractTableSolver: An abstract table solver type, which is not utilized in this context.

vs: A vector of values associated with the problem-solving process.

Returns

nothing: Indicates that no solution is produced due to the absence of a problem.

OptimalBranchingCore.subcovers — Method

subcovers(tbl::BranchingTable{INT}) where {INT}

Generates subcovers from a branching table.

Arguments

tbl::BranchingTable{INT}: The branching table containing bit strings.

Returns

Vector{SubCover{INT}}: A vector of SubCover objects generated from the branching table.

Description

This function calls the subcovers function with the bit length and table from the provided branching table to generate the corresponding subcovers.

OptimalBranchingCore.subcovers — Method

subcovers(n::Int, bss::AbstractVector{Vector{INT}}) where {INT}

Generates a set of subcovers from the given bit strings.

Arguments

n::Int: The length of the bit strings.

bss::AbstractVector{Vector{INT}}: A collection of vectors containing bit strings.

Returns

Vector{SubCover{INT}}: A vector of SubCover objects representing the generated subcovers.

Description

This function concatenates the input vectors of bit strings and iteratively generates clauses. It maintains a set of all unique clauses and uses a temporary list to explore new clauses formed by combining existing ones. The resulting subcovers are created based on the covered items for each clause.

OptimalBranchingCore.AbstractBranchingStrategy — Type

AbstractBranchingStrategy

An abstract type representing a branching strategy in the optimization process.

OptimalBranchingCore.AbstractMeasure — Type

AbstractMeasure

An abstract type representing a measure in the context of branching problems. This serves as a base type for all specific measure implementations.

OptimalBranchingCore.AbstractProblem — Type

AbstractProblem

An abstract type representing a generic problem in the optimal branching framework. This serves as a base type for all specific problem types that will be implemented.

Fields

None

OptimalBranchingCore.AbstractPruner — Type

AbstractPruner

An abstract type representing a pruner in the context of branching problems. This serves as a base type for all specific pruner implementations.

OptimalBranchingCore.AbstractReducer — Type

AbstractReducer

An abstract type representing a reducer in the context of branching problems. This serves as a base type for all specific reducer implementations.

OptimalBranchingCore.AbstractResult — Type

AbstractResult

An abstract type representing a generic result in the optimal branching framework. This serves as a base type for all specific result types that will be implemented.

Fields

None

OptimalBranchingCore.AbstractSelector — Type

AbstractSelector

An abstract type representing a selector in the context of branching problems. This serves as a base type for all specific selector implementations.

OptimalBranchingCore.AbstractSetCoverSolver — Type

AbstractSetCoverSolver

An abstract type representing a solver for set covering problems. This serves as a base type for all specific set cover solver implementations.

OptimalBranchingCore.AbstractTableSolver — Type

AbstractTableSolver

An abstract type representing a solver for table-based problems. This serves as a base type for all specific table solver implementations.

OptimalBranchingCore.Branch — Type

struct Branch

A struct representing a branching strategy.

Fields

vertices_removed::Vector{Int}: A vector of integers representing the vertices removed in the branching strategy.

mis::Int: An integer representing the maximum independent set (MIS) size of the branching strategy.

OptimalBranchingCore.BranchingTable — Type

BranchingTable{INT}

A table of branching configurations. The table is a vector of vectors of INT. Type parameters are:

INT: The number of integers as the storage.

Fields

bit_length::Int: The length of the bit string.

table::Vector{Vector{INT}}: The table of bitstrings used for branching.

To cover the branching table, at least one clause in each row must be satisfied.

OptimalBranchingCore.Clause — Type

Clause{INT <: Integer}

A Clause is conjunction of literals, which is specified by a pair of bit strings. The type parameter INT is the integer type for storing the bit strings.

Fields

mask: A bit string that indicates the variables involved in the clause.

val: A bit string that indicates the positive literals in the clause.

Examples

To check if a bit string satisfies a clause, use OptimalBranchingCore.covered_by.

julia> using OptimalBranchingCore

julia> clause = Clause(0b1110, 0b1010)
Clause{UInt8}: #2 ∧ ¬#3 ∧ #4

julia> OptimalBranchingCore.covered_by(0b1110, clause)
false

julia> OptimalBranchingCore.covered_by(0b1010, clause)
true

OptimalBranchingCore.DNF — Type

DNF{INT}

A data structure representing a Disjunctive Normal Form (DNF) expression. A DNF is a logical formula that is a disjunction of one or more conjunctions of literals.

Fields

clauses::Vector{Clause{INT}}: A vector of Clause objects representing the individual clauses in the DNF.

OptimalBranchingCore.IPSolver — Type

IPSolver

A struct representing an integer programming solver for set covering problems.

Fields

max_itr::Int: The maximum number of iterations allowed for the solver.

Constructors

IPSolver(max_itr::Int): Creates a new instance of IPSolver with a specified maximum number of iterations.

IPSolver(): Creates a new instance of IPSolver with a default maximum of 10 iterations.

OptimalBranchingCore.LPSolver — Type

LPSolver

A struct representing a linear programming solver for set covering problems.

Fields

max_itr::Int: The maximum number of iterations allowed for the solver.

Constructors

LPSolver(max_itr::Int): Creates a new instance of LPSolver with a specified maximum number of iterations.

LPSolver(): Creates a new instance of LPSolver with a default maximum of 10 iterations.

OptimalBranchingCore.NoProblem — Type

NoProblem

A concrete implementation of AbstractProblem representing the absence of a problem. This is used as a placeholder in scenarios where no valid problem is present.

Fields

None

OptimalBranchingCore.NoPruner — Type

NoPruner

A struct representing a no-operation pruner. This pruner does not modify the branching table during the pruning process.

OptimalBranchingCore.NoResult — Type

NoResult

A concrete implementation of AbstractResult representing the absence of a result. This is used as a placeholder in scenarios where no valid result is present.

Fields

None

OptimalBranchingCore.OptBranchingStrategy — Type

OptBranchingStrategy

A struct representing an optimal branching strategy that utilizes various components for solving optimization problems.

Fields

table_solver::TS: An instance of a table solver, which is responsible for solving the underlying table representation of the problem.

set_cover_solver::SCS: An instance of a set cover solver, which is used to solve the set covering problem.

pruner::PR: An instance of a pruner, which is used to eliminate unnecessary branches in the search space.

selector::SL: An instance of a selector, which is responsible for selecting the next branching variable or decision.

measure::M: An instance of a measure, which is used to evaluate the performance of the branching strategy.

OptimalBranchingCore.SolverConfig — Type

SolverConfig

A struct representing the configuration for a solver, including the reducer and branching strategy.

Fields

reducer::R: An instance of a reducer, which is responsible for reducing the problem size.

branching_strategy::B: An instance of a branching strategy, which guides the search process.

result_type::Type{TR}: The type of the result that the solver will produce.

OptimalBranchingCore.SubCover — Type

SubCover{INT <: Integer}

A subcover is a pair of a set of integers ids and a clause clause. The ids for the truth covered by the clause.

INT: The number of integers as the storage.