OptimalBranchingMIS

Documentation for the MIS module OptimalBranchingMIS.jl, which provides the maximum independent set solver based on the optimal branching algorithm.

API

OptimalBranchingMIS.D3Measure

OptimalBranchingMIS.EnvFilter

OptimalBranchingMIS.MISCount

OptimalBranchingMIS.MISProblem

OptimalBranchingMIS.MISReducer

OptimalBranchingMIS.MISSize

OptimalBranchingMIS.MinBoundarySelector

OptimalBranchingMIS.NumOfVertices

OptimalBranchingMIS.TensorNetworkSolver

OptimalBranchingCore.apply

OptimalBranchingCore.measure

OptimalBranchingCore.measure

OptimalBranchingCore.problem_reduce

OptimalBranchingCore.result

OptimalBranchingCore.select

OptimalBranchingCore.solve_table

OptimalBranchingMIS.closed_neighbors

OptimalBranchingMIS.counting_mis1

OptimalBranchingMIS.counting_mis2

OptimalBranchingMIS.counting_xiao2013

OptimalBranchingMIS.neighbor_cover

OptimalBranchingMIS.neighbors_2nd

OptimalBranchingMIS.open_neighbors

OptimalBranchingMIS.open_vertices

OptimalBranchingMIS.removed_vertices

OptimalBranchingCore.apply — Method

apply(p::MISProblem, clause::Clause{INT}, vertices::Vector{T}) where {INT<:Integer, T<:Integer}

Applies the given clause to the specified vertices of the MISProblem, removing the vertices that are affected by the clause.

Arguments

p::MISProblem: The problem instance containing the graph.

clause::Clause{INT}: The clause to be applied, which may affect the vertices.

vertices::Vector{T}: A vector of vertices to be considered for removal.

Returns

MISProblem: A new instance of MISProblem with the specified vertices removed from the graph.

OptimalBranchingCore.measure — Method

measure(p::MISProblem, ::D3Measure)

Calculates the D3 measure for the given MISProblem, which is defined as the sum of the maximum degree of each vertex minus 2, for all vertices in the graph.

Arguments

p::MISProblem: The problem instance containing the graph.

Returns

Int: The computed D3 measure value.

OptimalBranchingCore.measure — Method

measure(p::MISProblem, ::NumOfVertices)

Calculates the number of vertices in the given MISProblem.

Arguments

p::MISProblem: The problem instance containing the graph.

Returns

Int: The number of vertices in the graph.

OptimalBranchingCore.problem_reduce — Method

problem_reduce(p::MISProblem, ::MISReducer, TR::Type{R}) where R<:AbstractResult

Reduces the given MISProblem by removing vertices based on their degrees and returns a new MISProblem instance along with the count of removed vertices.

Arguments

p::MISProblem: The problem instance containing the graph to be reduced.

::MISReducer: An instance of the MISReducer struct.

TR::Type{R}: The type of the result expected.

Returns

A tuple containing:

A new MISProblem instance with specified vertices removed.

An integer representing the count of removed vertices.

Description

The function checks the number of vertices in the graph:

If there are no vertices, it returns a NoProblem instance and a count of 0.

If there is one vertex, it returns a NoProblem instance and a count of 1.

If there are two vertices, it returns a NoProblem instance and a count based on the presence of an edge between them.

For graphs with more than two vertices, it calculates the degrees of the vertices and identifies the vertex with the minimum degree to determine which vertices to remove.

OptimalBranchingCore.result — Method

result(p::MISProblem, clause::Clause{INT}, vertices::Vector{T}, TR::Type{R}) where {INT<:Integer, R<:AbstractResult, T<:Integer}

Calculates the result of applying the given clause to the specified vertices in the MISProblem.

Arguments

p::MISProblem: The problem instance containing the graph.

clause::Clause{INT}: The clause whose value is to be counted.

vertices::Vector{T}: A vector of vertices to be considered (not used in the calculation).

TR::Type{R}: The type of the result expected.

Returns

Int: The count of ones in the clause's value.

OptimalBranchingCore.select — Method

select(p::MISProblem, m::M, selector::MinBoundarySelector)

Selects a subgraph from the given MISProblem based on the minimum number of open vertices within k-layers of neighbors.

Arguments

p::MISProblem: The problem instance containing the graph from which to select vertices.

m::M: An instance of a measure associated with the problem (not used in this function).

selector::MinBoundarySelector: The selector strategy that defines the number of neighbor layers to consider.

Returns

Vector{Int}: A vector of vertices representing the selected subgraph with the minimum number of open vertices.

Description

This function iterates through each vertex in the graph, computes the neighbor cover for each vertex up to k layers, and selects the vertex that results in the minimum number of open vertices. The selected vertices are returned as a vector.

OptimalBranchingCore.solve_table — Method

solve_table(p::MISProblem, solver::TensorNetworkSolver, vs::Vector{Int})

Calculates the reduced alpha configurations for a given Maximum Independent Set (MIS) problem.

Arguments

p::MISProblem: The problem instance containing the graph and vertices.

solver::TensorNetworkSolver: The solver used for optimizing the tensor network.

vs::Vector{Int}: A vector of vertices to be considered in the subgraph.

Returns

BranchingTable: A table containing the reduced alpha configurations derived from the induced subgraph.

Description

This function first identifies the open vertices in the graph associated with the given problem. It then creates an induced subgraph based on the specified vertices and computes the reduced alpha configurations using the provided tensor network solver.

OptimalBranchingMIS.closed_neighbors — Method

closed_neighbors(g::SimpleGraph, vertices::Vector{Int})

Returns a set of vertices that includes the input vertices as well as their open neighbors.

Arguments

g::SimpleGraph: The input graph.

vertices::Vector{Int}: The vertices for which closed neighbors are to be computed.

Returns

A set of vertices that includes the input vertices as well as their open neighbors.

OptimalBranchingMIS.counting_mis1 — Method

counting_mis1(g::SimpleGraph)

Calculates the maximum independent set (MIS) for a given SimpleGraph by first converting it into an EliminateGraph and then applying the counting_mis1 function.

Arguments

g::SimpleGraph: The input graph for which the maximum independent set is to be calculated.

Returns

MISCount: The size of the maximum independent set for the provided graph and the count of branches.

OptimalBranchingMIS.counting_mis2 — Method

counting_mis2(g::SimpleGraph)

Calculates the size of the maximum independent set and the count of branches for a given SimpleGraph.

OptimalBranchingMIS.counting_xiao2013 — Method

counting_xiao2013(g::SimpleGraph)

This function counts the maximum independent set (MIS) in a given simple graph using the Xiao 2013 algorithm.

Arguments

g::SimpleGraph: A simple graph for which the maximum independent set is to be counted.

Returns

CountingMIS: An object representing the size of the maximum independent set and the count of branches.

OptimalBranchingMIS.neighbor_cover — Method

neighbor_cover(g::SimpleGraph, v::Int, k::Int)

Compute the neighbor cover of a vertex in a graph.

Arguments

g::SimpleGraph: The input graph.

v::Int: The vertex for which to compute the neighbor cover.

k::Int: The number of iterations to perform.

Returns

vertices: An array containing the vertices in the neighbor cover.

openvertices: An array containing the open vertices in the neighbor cover.

OptimalBranchingMIS.neighbors_2nd — Method

neighbors_2nd(g::SimpleGraph, v::Int)

Return the second-order neighbors of a vertex v in a simple graph g.

Arguments

g::SimpleGraph: The simple graph.

v::Int: The vertex.

Returns

Array{Int}: An array of second-order neighbors of v.

OptimalBranchingMIS.open_neighbors — Method

open_neighbors(g::SimpleGraph, vertices::Vector{Int})

Returns a vector of vertices in the graph g, which are neighbors of the given vertices and not in the given vertices.

Arguments

g::SimpleGraph: The graph in which to find the open neighbors.

vertices::Vector{Int}: The vertices for which to find the open neighbors.

Returns

A vector of open neighbors of the given vertices.

OptimalBranchingMIS.open_vertices — Method

open_vertices(g::SimpleGraph, vertices::Vector{Int})

Remove vertices from the given vector that are connected to all other vertices in the graph.

Arguments

g::SimpleGraph: The graph object.

vertices::Vector{Int}: The vector of vertices.

Returns

Vector{Int}: The open vertices.

OptimalBranchingMIS.removed_vertices — Method

removed_vertices(vertices::Vector{Int}, g::SimpleGraph, clause::Clause{N}) where N

Given a list of vertices, a graph, and a clause, this function returns a list of removed vertices.

The vertices argument is a vector of integers representing the vertices to consider. The g argument is a SimpleGraph object representing the graph. The clause argument is a Clause object representing a clause.

The function iterates over the vertices and checks if the corresponding bit in the clause.mask is 1. If it is, the vertex is added to the list of removed vertices (rvs). If the corresponding bit in the clause.val is also 1, the neighbors of the vertex are also added to rvs.

The function returns the list of removed vertices with duplicates removed.

OptimalBranchingMIS.D3Measure — Type

D3Measure

A struct representing a measure that calculates the sum of the maximum degree minus 2 for each vertex in the graph.

Fields

None

OptimalBranchingMIS.EnvFilter — Type

EnvFilter

A struct representing an environment filter used in the context of pruning branching tables. This filter serves as a specific implementation of the AbstractPruner type.

OptimalBranchingMIS.MISCount — Type

struct MISCount <: AbstractResult

Represents the count of Maximum Independent Sets (MIS).

Fields

mis_size::Int: The size of the Maximum Independent Set.

mis_count::Int: The number of Maximum Independent Sets of that size.

Constructors

MISCount(mis_size::Int): Creates a MISCount with the given size and initializes the count to 1.

MISCount(mis_size::Int, mis_count::Int): Creates a MISCount with the specified size and count.

OptimalBranchingMIS.MISProblem — Type

mutable struct MISProblem <: AbstractProblem

Represents a Maximum Independent Set (MIS) problem.

Fields

g::SimpleGraph: The graph associated with the MIS problem.

Methods

copy(p::MISProblem): Creates a copy of the given MISProblem.

Base.show(io::IO, p::MISProblem): Displays the number of vertices in the MISProblem.

OptimalBranchingMIS.MISReducer — Type

MISReducer

A struct representing a reducer for the Maximum Independent Set (MIS) problem. This struct serves as a specific implementation of the AbstractReducer type.

OptimalBranchingMIS.MISSize — Type

struct MISSize <: AbstractResult

Represents the size of a Maximum Independent Set (MIS).

Fields

mis_size::Int: The size of the Maximum Independent Set.

OptimalBranchingMIS.MinBoundarySelector — Type

struct MinBoundarySelector <: AbstractVertexSelector

The MinBoundarySelector struct represents a strategy for selecting a subgraph with the minimum number of open vertices by k-layers of neighbors.

Fields

k::Int: The number of layers of neighbors to consider when selecting the subgraph.

OptimalBranchingMIS.NumOfVertices — Type

NumOfVertices

A struct representing a measure that counts the number of vertices in a graph. Each vertex is counted as 1.

Fields

None

OptimalBranchingMIS.TensorNetworkSolver — Type

TensorNetworkSolver

A struct representing a solver for tensor network problems. This struct serves as a specific implementation of the AbstractTableSolver type.